backend/modules/ai_research_analyst/requirements.txt
# AI Research Analyst Dependencies

# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
requests>=2.28.0
tqdm>=4.64.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
lxml>=4.9.0  # For better XML parsing if needed

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0 